feat: 러닝 완료 시 기록 저장과 AI 호출 연동, 러닝 세팅 반환하는 컨트롤러 추가

+++ src/main/java/com/example/runity/domain/RealTimeRunning.java
@@ -70,7 +70,7 @@ public class RealTimeRunning {
     private String feedback_advice;
 
     @Column
-    private Float feedback_earlySpeedDeviation;
+    private Double feedback_earlySpeedDeviation;
 
     @Column
     private Integer focusScore;
